pyproject.toml
[project]
name = "advancedagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "firecrawl-py>=2.8.0",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
]
